% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{inbredPermute}
\alias{inbredPermute}
\alias{inbredPermute-package}
\title{Simple permutation procedures for distributions of rxy values amongst parents}
\description{
\code{inbredPermute} is a package.
Add more description here.
}

