% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_kingroup}
\alias{read_kingroup}
\title{Read in the R_xy values from KinGroup}
\usage{
read_kingroup(file, ...)
}
\arguments{
\item{file}{The KinGroup output file.}

\item{...}{Other options to pass to read.table.  One possibility would
be \code{dec = ","} for dealing with European number styles that use a
comma for the decimal point.}
}
\description{
This assumes that the output is the lower triangle
of the values, in tab-delimited format, with the
diagonal and upper triangular values present in the
file (i.e. the TABS are there) but simply nothings, or
single spaces.
}
\examples{
rxy <- read_kingroup(system.file("data_files/MKHM_rxy.txt", package="inbredPermute", mustWork = T), dec=",")
}

