% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_kingroup_csv}
\alias{read_kingroup_csv}
\title{Read in kingroup results from the raw kingroup CSV format}
\usage{
read_kingroup_csv(file, ItalianCommas = FALSE)
}
\arguments{
\item{file}{The Kingroup CSV output file.   Note: you cannot have spaces in individual
IDs and expect this to work.}

\item{ItalianCommas}{Logical.  True if we need to convert decimal commas to decimal periods.
Named this in honor of our delightful Italian visitor David who had crazy commas in his numbers!}
}
\description{
This assumes that the file is just a CSV file that Kingroup produced
of Rxy values.
}
\examples{
rxy <- read_kingroup_csv(system.file("data_files/MKHM_kingroup_output.csv", package="inbredPermute", mustWork = T), ItalianCommas = TRUE)
}

